Bug fix for regional grids.
Fix bug where a regonal grid that is asymmetric (i.e. it does not have the same number of cells in the x and y diretions, and/or it is not symmetric when reflected through its centerline at i=nx/2 or j=ny/2) will have its dy values set incorrectly or have its indices go out of array bounds.